Spectrum injector
This flashable zip creates user configureable profiles in /data/spectrum/. Users can add profiles in the the dummy files there that will be loaded by the Spectrum app. Boot block detection was taken from lazyflasher by @jcadduono, so huge thanks to him for that. This overall idea was taken from lkt and lambda kernel by @GalaticStryder, so thanks to him as well.

Signed-off-by: Joe Maples <joe@frap129.org>